test(rawdata): keep PR796/797 regressions in RawDataParserTest only
